feat/table: refact: extract settings into component
